blockverify/requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
google-generativeai>=0.3.0
